[#15] refactor: kakao 리다이렉트 주소 수정 및 디렉토리 구조 변경
