+++ src/main/java/com/itoxi/petnuri/domain/member/dto/response/KakaoToken.java
@@ -1,4 +1,4 @@
-package com.itoxi.petnuri.domain.member.dto;
+package com.itoxi.petnuri.domain.member.dto.response;
 
 import lombok.Getter;
